Created WA GeoJSON
This is taking data from source and changing it to be common name suburb_name all null fields have been removed as they were not needed for the project.

We have kept created dates and other date fields as we will do our best to keep this updated in the near future.

If you see an error with the dataset please open a issue report.

 https://data.gov.au/data/dataset/wa-suburb-locality-boundaries-geoscape-administrative-boundaries